requirements.txt
fastapi
uvicorn
streamlit

langchain
langchain-community
langchain-core

chromadb

sentence-transformers
huggingface-hub

llama-index-core
llama-index-embeddings-huggingface
llama-index-llms-ollama

unstructured
python-docx
pymupdf
